Refactor RequestBuilder & request stream reading into separate class
