+++ src/test/java/com/eighthlight/fabrial/test/file/FileHttpResponderTest.java
@@ -59,19 +59,22 @@ public class FileHttpResponderTest {
           allFilePaths.addAll(existingFilePaths);
           allFilePaths.addAll(nonExistingFilePaths);
           allFilePaths.forEach(p -> {
-            URI pathURI;
-            try {
-              pathURI = new URI(p.toString());
-            } catch (URISyntaxException e) {
-              throw new RuntimeException(e);
-            }
-            Request req = new Request(HttpVersion.ONE_ONE, Method.HEAD, pathURI);
+            Request req = new Request(HttpVersion.ONE_ONE, Method.HEAD, makeURI(p));
             int expectedStatus = existingFilePaths.contains(p) ? 200 : 404;
             assertThat(
                 responder.getResponse(req),
                 equalTo(new Response(HttpVersion.ONE_ONE, expectedStatus, null)));
           });
         });
+  }
+
+  private URI makeURI(Path p) {
+    URI pathURI;
+    try {
+      return new URI(p.toString());
+    } catch (URISyntaxException e) {
+      throw new RuntimeException(e);
+    }
   }
 
   @Test
@@ -82,13 +85,7 @@ public class FileHttpResponderTest {
           Set<Path> nonExistingFilePaths = files.get(1);
           FileHttpResponder responder = responderForListOfExistingFiles(existingFilePaths);
           nonExistingFilePaths.forEach(p -> {
-            URI pathURI;
-            try {
-              pathURI = new URI(p.toString());
-            } catch (URISyntaxException e) {
-              throw new RuntimeException(e);
-            }
-            Request req = new Request(HttpVersion.ONE_ONE, Method.DELETE, pathURI);
+            Request req = new Request(HttpVersion.ONE_ONE, Method.DELETE, makeURI(p));
             assertThat(
                 responder.getResponse(req),
                 equalTo(new Response(HttpVersion.ONE_ONE, 501, null)));
